requirements.txt
requests
py-cpuinfo
